Refactor spec helper methods to their own module
