+++ spec/spec_helper.rb
@@ -6,6 +6,8 @@ SimpleCov.start
 
 require 'json'
 
+require_relative 'helper_methods'
+
 # This file was generated by the `rspec --init` command. Conventionally, all
 # specs live under a `spec` directory, which RSpec adds to the `$LOAD_PATH`.
 # The generated `.rspec` file contains `--require spec_helper` which will cause
@@ -77,6 +79,9 @@ RSpec.configure do |config|
   Kernel.srand config.seed
 
   config.filter_run_when_matching focus: true
+
+  config.extend(HelperMethods)
+  config.include(HelperMethods)
 end
 
 # Keeps track of what events are run; if one isn't run it will fail
@@ -110,41 +115,3 @@ RSpec::Matchers.define :something_not_including do |x|
 end
 
 RSpec::Matchers.define_negated_matcher :an_array_excluding, :include
-
-def load_data_file(*name)
-  JSON.parse(File.read("#{File.dirname(__FILE__)}/json_examples/#{name.join('/')}.json"))
-end
-
-# Creates a helper method that gives access to a particular fixture's data.
-# @example Load the JSON file at "spec/data/folder/filename.json" as a "data_name" helper method
-#   fixture :data_name, [:folder, :filename]
-# @param name [Symbol] The name the helper method should have
-# @param path [Array<Symbol>] The path to the data file to load, originating from "spec/data"
-def fixture(name, path)
-  let name do
-    load_data_file(*path)
-  end
-end
-
-# Creates a helper method that gives access to a specific property on a particular fixture.
-# @example Add a helper method called "property_value" for `data_name['properties'][0]['value'].to_i`
-#   fixture_property :property_value, :data_name, ['properties', 0, 'value'], :to_i
-# @param name [Symbol] The name the helper method should have
-# @param fixture [Symbol] The name of the fixture the property is on
-# @param trace [Array] The objects to consecutively pass to the #[] method when accessing the data.
-# @param filter [Symbol, nil] An optional method to call on the result, in case some conversion is necessary.
-def fixture_property(name, fixture, trace, filter = nil)
-  let name do
-    data = send(fixture)
-
-    trace.each do |e|
-      data = data[e]
-    end
-
-    if filter
-      data.send(filter)
-    else
-      data
-    end
-  end
-end
